+++ test/functional/abc_p2p_avalanche_proof_voting.py
@@ -50,6 +50,7 @@ class AvalancheProofVotingTest(BitcoinTestFramework):
                 "-avaminquorumstake=0",
                 "-avaminavaproofsnodecount=0",
                 "-whitelist=noban@127.0.0.1",
+                "-persistavapeers=0",
             ],
         ]
         self.supports_cli = False
@@ -328,6 +329,7 @@ class AvalancheProofVotingTest(BitcoinTestFramework):
                 "-avaminquorumstake=0",
                 "-avaminavaproofsnodecount=0",
                 "-whitelist=noban@127.0.0.1",
+                "-persistavapeers=0",
             ],
         )
 
@@ -537,6 +539,7 @@ class AvalancheProofVotingTest(BitcoinTestFramework):
                 "-avaminquorumstake=0",
                 "-avaminavaproofsnodecount=0",
                 "-whitelist=noban@127.0.0.1",
+                "-persistavapeers=0",
             ],
         )