+++ test/functional/abc_p2p_proof_inventory.py
@@ -61,6 +61,7 @@ class ProofInventoryTest(BitcoinTestFramework):
                 "-avaproofstakeutxoconfirmations=2",
                 "-avacooldown=0",
                 "-whitelist=noban@127.0.0.1",
+                "-persistavapeers=0",
             ]
         ] * self.num_nodes